journify-3.2.11

+++ src/app/mytrips/page.tsx
@@ -1,43 +1,33 @@
 
 'use client'
 import TripPageCard from "@/component/TripPageCard";
-import { useEffect, useState } from "react";
+import { SetStateAction, useEffect, useState } from "react";
 import { useRouter } from 'next/navigation';
 import { IoMdAdd } from "react-icons/io";
 import "react-day-picker/style.css";
 import { auth } from '@/lib/firebase';
-import { addDoc, collection, query, onSnapshot, doc, deleteDoc, updateDoc } from "firebase/firestore";
+import { addDoc, collection, query, onSnapshot, doc, deleteDoc, updateDoc, getDoc, increment, setDoc } from "firebase/firestore";
 import { db } from "@/lib/firebase";
 import { useAuth } from '@/context/AuthContext';
 import { Timestamp } from "firebase/firestore";
 import CreateTrip from "@/component/CreateTrip";
+import UpdateTrip from "@/component/UpdateTrip";
+import { Country, TripDaySchedule } from "../type/trip";
 
-interface TripTime {
-    tripFrom: Date;
-    tripTo: Date;
-}
+
 interface Trip {
     id?: string;
     tripName: string;
-    person: Number;
-    tripTime: TripTime;
-    isPublic: boolean;
-    tripCountry: string;
-    createAt: Timestamp;
-    updateAt: Timestamp;
-}
-interface FirestoreTripTime {
-    tripFrom: Timestamp;
-    tripTo: Timestamp;
-}
-interface FirestoreTrip {
-    tripName: string;
     person: number;
-    tripTime: FirestoreTripTime;
+    tripTime: {
+        tripFrom: Timestamp;
+        tripTo: Timestamp;
+    };
     isPublic: boolean;
-    tripCountry: string;
+    tripCountry: Country[];
     createAt: Timestamp;
     updateAt: Timestamp;
+    tripDaySchedule?: TripDaySchedule[] | null;
 }
 
 export default function MyTrips() {
@@ -53,8 +43,15 @@ export default function MyTrips() {
     // 建立旅程狀態
     const [isAddTrip, setIsAddTrip] = useState<boolean>(false);
 
+    // 更新旅程狀態
+    const [isEditingTrip, setIsEditingTrip] = useState<boolean>(false);
+    const [editTripData, setEditTripData] = useState<Trip | null>(null);
+
     // 使用者資料庫的旅程資料
     const [trips, setTrips] = useState<Trip[]>([]);
+
+    // 儲存旅程資料
+    const [saveStatus, setSaveStatus] = useState<"idle" | "saving" | "success" | "error">("idle");
 
     // 使用者是否為登入狀態
     useEffect(() => {
@@ -73,16 +70,16 @@ export default function MyTrips() {
 
         const unsubscribe = onSnapshot(q, (snapshot) => {
             const data: Trip[] = snapshot.docs.map((doc) => {
-                const tripData = doc.data() as FirestoreTrip;
-                const tripTime = {
-                    tripFrom: tripData.tripTime.tripFrom.toDate(),
-                    tripTo: tripData.tripTime.tripTo.toDate(),
-                };
+                const tripData = doc.data() as Trip;
+                // const tripTime = {
+                //     tripFrom: tripData.tripTime.tripFrom.toDate(),
+                //     tripTo: tripData.tripTime.tripTo.toDate(),
+                // };
                 return {
                     id: doc.id,
                     tripName: tripData.tripName,
                     person: tripData.person,
-                    tripTime,
+                    tripTime: tripData.tripTime,
                     isPublic: tripData.isPublic,
                     tripCountry: tripData.tripCountry,
                     createAt: tripData.createAt,
@@ -91,7 +88,7 @@ export default function MyTrips() {
             });
             setTrips(data);
             setIsloading(false);
-
+            console.log(data);
         });
         return () => unsubscribe();
     }, [user?.uid]);
@@ -99,11 +96,14 @@ export default function MyTrips() {
     // 刪除旅程
     async function deleteTrip(userId: string | undefined, tripId: string) {
         if (userId && tripId) {
+            const tripCountries = trips.find(item => item.id === tripId)?.tripCountry;
+            if (!tripCountries) return;
             try {
                 const tripRef = doc(db, "users", userId, "trips", tripId);
                 const allTripRef = doc(db, "all_trips", tripId);
                 await deleteDoc(tripRef);
                 await deleteDoc(allTripRef);
+                updateCountryStatsOnDelete(tripCountries);
                 console.log("Trip deleted successfully");
             } catch (error) {
                 console.error("Failed to delete trip:", error);
@@ -132,6 +132,83 @@ export default function MyTrips() {
         }
     }
 
+    /**
+    * 建立旅程時，旅遊國家的統計資料表內該國家數字+1
+    * @param {Country[]} tripCountry - 要建立的國家list
+    */
+    async function updateCountryStatsOnCreate(tripCountry: Country[]) {
+        const promises = tripCountry.map(async (country) => {
+            const countryRef = doc(db, "countryStats", country.countryCode);
+            const countrySnap = await getDoc(countryRef);
+
+            if (countrySnap.exists()) {
+                // 已存在，count +1
+                await updateDoc(countryRef, {
+                    count: increment(1)
+                });
+            } else {
+                // 不存在，新增該國家統計資料
+                await setDoc(countryRef, {
+                    code: country.countryCode,
+                    name: country.countryName,
+                    count: 1
+                });
+            }
+        });
+
+        await Promise.all(promises);
+    }
+
+    /**
+    * 刪除旅程時，旅遊國家的統計資料表內該國家數字-1
+    * @param {Country[]} tripCountry - 要建立的國家list
+    */
+    async function updateCountryStatsOnDelete(tripCountry: Country[]) {
+        const promises = tripCountry.map(async (country) => {
+            const countryRef = doc(db, "countryStats", country.countryCode);
+            const countrySnap = await getDoc(countryRef);
+
+            if (countrySnap.exists()) {
+                const currentCount = countrySnap.data().count || 0;
+
+                if (currentCount > 1) {
+                    //  不是最後一筆，直接扣 1
+                    await updateDoc(countryRef, {
+                        count: increment(-1)
+                    });
+                } else {
+                    //  如果是最後一筆就直接刪掉這筆統計
+                    await setDoc(countryRef, {
+                        code: country.countryCode,
+                        name: country.countryName,
+                        count: 0
+                    });
+                }
+            }
+        });
+
+        await Promise.all(promises);
+    }
+
+    /**
+    * 更新旅程時，旅遊國家的統計資料表內，移除的國家數字-1，新增的國家數字+1
+    * @param {Country[]} oldCountries - 舊的國家list
+    * @param {Country[]} newCountries - 更新的國家list
+    */
+    async function updateCountryStatsOnEdit(
+        oldCountries: Country[],
+        newCountries: Country[]
+    ) {
+        const oldSet = new Set(oldCountries.map(c => c.countryCode));
+        const newSet = new Set(newCountries.map(c => c.countryCode));
+
+        const toAdd = newCountries.filter(c => !oldSet.has(c.countryCode));
+        const toRemove = oldCountries.filter(c => !newSet.has(c.countryCode));
+
+        await updateCountryStatsOnCreate(toAdd);   // 新增的加一
+        await updateCountryStatsOnDelete(toRemove); // 移除的減一
+    }
+
 
     return (
         <div className="w-full h-full ">
@@ -141,13 +218,24 @@ export default function MyTrips() {
                     <p className="text-mywhite-100">旅雀加載中...請稍後</p>
                 </div>
             }
+            {saveStatus !== "idle" && (
+                <div className='fixed top-0 w-full h-full bg-myzinc900-60 z-1000 flex flex-col items-center justify-center'>
+                    <img src="/loading.gif" className="w-30 h-30 " />
+                    <div className='w-fit h-fit px-5 py-3  text-mywhite-100 text-base-500'>
+                        {saveStatus === "saving" && <span >儲存中...</span>}
+                        {saveStatus === "success" && <span >儲存成功！</span>}
+                        {saveStatus === "error" && <span >儲存失敗，請稍後再試</span>}
+                    </div>
+                </div>)}
+
             <div className="w-full h-fit flex flex-col p-10 mb-20">
                 <div className="w-fit h-fit mb-6">
                     <p className="text-2xl font-bold text-myblue-800">我的旅程</p>
                 </div>
                 <div id="tripsWrapper" className="grid gap-8 grid-cols-1 sm:grid-cols-2 md:grid-cols-3 w-full lg:grid-cols-4">
                     {/* trip card */}
-                    {trips.map((item) => (<TripPageCard key={item.id} item={item} tripPerson={item.person} deleteTrip={deleteTrip} userId={userId} updateTripPrivate={updateTripPrivate} />))}
+                    {trips.map((item) => (<TripPageCard key={item.id} item={item} tripPerson={item.person} deleteTrip={deleteTrip} userId={userId} updateTripPrivate={updateTripPrivate}
+                        setIsEditingTrip={setIsEditingTrip} setEditTripData={setEditTripData} />))}
                 </div>
                 <button className="fixed bottom-6 right-10 w-30 h-10 bg-primary-300 ml-auto 
                 rounded-full text-base text-myblue-600 font-bold flex items-center 
@@ -157,7 +245,8 @@ export default function MyTrips() {
                     建立旅程
                 </button>
             </div>
-            {isAddTrip && <CreateTrip userId={userId} setIsAddTrip={setIsAddTrip} />}
+            {isAddTrip && <CreateTrip userId={userId} setIsAddTrip={setIsAddTrip} updateCountryStatsOnCreate={updateCountryStatsOnCreate} />}
+            {isEditingTrip && <UpdateTrip userId={userId} setIsEditingTrip={setIsEditingTrip} editTripData={editTripData} setSaveStatus={setSaveStatus} updateCountryStatsOnEdit={updateCountryStatsOnEdit} />}
         </div>
     )
 }
